requirements.txt
jira
pyyaml
openai